Ecommerce/requirements.txt
Django==2.0.7
djangorestframework==3.8.2
mysql-connector==2.1.6
Pillow==5.2.0
pytz==2018.5
